requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# LLM & LangChain
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
google-generativeai>=0.5.0

# RAG & Vector Store
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# ASR
openai-whisper>=20230918

# Math Tools
sympy>=1.12
numpy>=1.24.0,<2.0.0
scipy>=1.11.0

# Config & Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Build tools (needed for whisper on older Python)
setuptools>=68.0.0
